+++ src/Component/AbstractPath.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Rowbot\URL\Component;
 
 use Rowbot\URL\String\Exception\UndefinedIndexException;
-use Rowbot\URL\URLRecord;
 
 use function count;
 
@@ -37,23 +36,6 @@ abstract class AbstractPath implements PathInterface
     public function isEmpty(): bool
     {
         return $this->list === [];
-    }
-
-    public function potentiallyStripTrailingSpaces(URLRecord $url): void
-    {
-        if (!$this->isOpaque()) {
-            return;
-        }
-
-        if ($url->fragment !== null) {
-            return;
-        }
-
-        if ($url->query !== null) {
-            return;
-        }
-
-        $this->list[0]->stripTrailingSpaces();
     }
 
     public function __clone()