+++ src/Component/PathInterface.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Rowbot\URL\Component;
 
 use Countable;
-use Rowbot\URL\URLRecord;
 use Stringable;
 
 interface PathInterface extends Countable, Stringable
@@ -22,9 +21,4 @@ interface PathInterface extends Countable, Stringable
      * @see https://url.spec.whatwg.org/#shorten-a-urls-path
      */
     public function shorten(Scheme $scheme): void;
-
-    /**
-     * @see https://url.spec.whatwg.org/#potentially-strip-trailing-spaces-from-an-opaque-path
-     */
-    public function potentiallyStripTrailingSpaces(URLRecord $url): void;
 }