+++ src/State/OpaquePathState.php
@@ -36,30 +36,42 @@ class OpaquePathState implements State
                 break;
             }
 
-            // 3. Otherwise:
-            // 3.1. If c is not the EOF code point, not a URL code point, and not U+0025 (%), validation error.
-            if ($codePoint !== CodePoint::EOF && $codePoint !== '%' && !CodePoint::isUrlCodePoint($codePoint)) {
-                // Validation error.
-                $context->logger?->notice('invalid-URL-unit', [
-                    'input'  => (string) $context->input,
-                    'column' => $context->iter->key() + 1,
-                ]);
+            // 3. Otherwise, if c is U+0020 SPACE:
+            if ($codePoint === "\x20") {
+                // 3.1. If remaining starts with U+003F (?) or U+003F (#), then append "%20" to url’s path.
+                // 3.2. Otherwise, append U+0020 SPACE to url’s path.
+                $remaining = $context->input->substr($context->iter->key() + 1);
+                $appendChar = "\x20";
 
-            // 3.2. If c is U+0025 (%) and remaining does not start with two ASCII hex digits, validation error.
-            } elseif (
-                $codePoint === '%'
-                && !$context->input->substr($context->iter->key() + 1)->startsWithTwoAsciiHexDigits()
-            ) {
-                // Validation error.
-                $context->logger?->notice('invalid-URL-unit', [
-                    'input'  => (string) $context->input,
-                    'column' => $context->iter->key() + 1,
-                ]);
-            }
+                if ($remaining->startsWith('?') || $remaining->startsWith('#')) {
+                    $appendChar = '%20';
+                }
 
-            // 3.3. If c is not the EOF code point, UTF-8 percent-encode c using the C0 control percent-encode set and
-            // append the result to url’s path.
-            if ($codePoint !== CodePoint::EOF) {
+                $context->url->path->first()->append($appendChar);
+
+            // 4. Otherwise, if c is not the EOF code point:
+            } elseif ($codePoint !== CodePoint::EOF) {
+                // 4.1. If c is not a URL code point and not U+0025 (%), invalid-URL-unit validation error.
+                if ($codePoint !== '%' && !CodePoint::isUrlCodePoint($codePoint)) {
+                    // Validation error.
+                    $context->logger?->notice('invalid-URL-unit', [
+                        'input'  => (string) $context->input,
+                        'column' => $context->iter->key() + 1,
+                    ]);
+
+                // 4.2. If c is U+0025 (%) and remaining does not start with two ASCII hex digits, invalid-URL-unit validation error.
+                } elseif (
+                    $codePoint === '%'
+                    && !$context->input->substr($context->iter->key() + 1)->startsWithTwoAsciiHexDigits()
+                ) {
+                    // Validation error.
+                    $context->logger?->notice('invalid-URL-unit', [
+                        'input'  => (string) $context->input,
+                        'column' => $context->iter->key() + 1,
+                    ]);
+                }
+
+                // 4.3. UTF-8 percent-encode c using the C0 control percent-encode set and append the result to url’s path.
                 $percentEncoder ??= new PercentEncoder();
                 $context->url->path->first()->append($percentEncoder->percentEncodeAfterEncoding(
                     'utf-8',