+++ src/URLSearchParams.php
@@ -287,10 +287,6 @@ class URLSearchParams implements Countable, Iterator, Stringable
         }
 
         $this->url->query = $query;
-
-        if ($query === null) {
-            $this->url->path->potentiallyStripTrailingSpaces($this->url);
-        }
     }
 
     /**